[ConstraintSystem] NFC: Extract `PotentialBindings` and auxiliary struct from `ConstraintSystem`
This opens up a posibility of using `PotentialBindings`
in `ConstraintGraphNode` and other places in `ConstraintGraph`.
